divide sidebar and about me page into their own reusable components
