+++ src/ui/CodeBlockDescription.tsx
@@ -1,0 +1,13 @@
+function CodeBlockDescription() {
+  return (
+    <div className='mt-8 flex justify-between border-t border-[#1E2D3D] py-6'>
+      <p className='w-5/6 text-base  text-[#607b96] sm:text-lg'>
+        My work here was 5 months ago. It was for the project
+        called “...”. Some other text can be placed here.
+      </p>
+      <span className='text-3xl sm:text-4xl'>&times;</span>
+    </div>
+  );
+}
+
+export default CodeBlockDescription;
