+++ src/ui/CodeBlocksList.tsx
@@ -1,0 +1,12 @@
+import CodeBlock from './CodeBlock';
+
+function CodeBlocksList() {
+  return (
+    <div className='px-0 md:px-6'>
+      <CodeBlock />
+      <CodeBlock />
+    </div>
+  );
+}
+
+export default CodeBlocksList;
