+++ src/ui/SidebarListItem.tsx
@@ -1,0 +1,57 @@
+function SidebarListItem() {
+  return (
+    <li className='cursor-pointer'>
+      <svg
+        width='22'
+        height='23'
+        viewBox='0 0 24 25'
+        fill='none'
+        xmlns='http://www.w3.org/2000/svg'
+      >
+        <g clip-path='url(#clip0_64_2227)'>
+          <path
+            d='M5 20.6519H19V22.6519H5V20.6519ZM12 18.6519C9.87827 18.6519 7.84344 17.809 6.34315 16.3087C4.84285 14.8084 4 12.7736 4 10.6519C4 8.53012 4.84285 6.49529 6.34315 4.995C7.84344 3.49471 9.87827 2.65186 12 2.65186C14.1217 2.65186 16.1566 3.49471 17.6569 4.995C19.1571 6.49529 20 8.53012 20 10.6519C20 12.7736 19.1571 14.8084 17.6569 16.3087C16.1566 17.809 14.1217 18.6519 12 18.6519Z'
+            fill='#607B96'
+          />
+        </g>
+        <defs>
+          <clipPath id='clip0_64_2227'>
+            <rect
+              width='24'
+              height='24'
+              fill='white'
+              transform='translate(0 0.651855)'
+            />
+          </clipPath>
+        </defs>
+      </svg>
+    </li>
+  );
+}
+
+export default SidebarListItem;
+
+// <svg
+//   width='22'
+//   height='23'
+//   viewBox='0 0 24 25'
+//   fill='none'
+//   xmlns='http://www.w3.org/2000/svg'
+// >
+//   <g opacity='0.4' clip-path='url(#clip0_64_2230)'>
+//     <path
+//       d='M3 3.65186H21C21.2652 3.65186 21.5196 3.75721 21.7071 3.94475C21.8946 4.13228 22 4.38664 22 4.65186V20.6519C22 20.9171 21.8946 21.1714 21.7071 21.359C21.5196 21.5465 21.2652 21.6519 21 21.6519H3C2.73478 21.6519 2.48043 21.5465 2.29289 21.359C2.10536 21.1714 2 20.9171 2 20.6519V4.65186C2 4.38664 2.10536 4.13228 2.29289 3.94475C2.48043 3.75721 2.73478 3.65186 3 3.65186ZM12 15.6519V17.6519H18V15.6519H12ZM8.414 12.6519L5.586 15.4799L7 16.8949L11.243 12.6519L7 8.40886L5.586 9.82386L8.414 12.6519Z'
+//       fill='#607B96'
+//     />
+//   </g>
+//   <defs>
+//     <clipPath id='clip0_64_2230'>
+//       <rect
+//         width='24'
+//         height='24'
+//         fill='white'
+//         transform='translate(0 0.651855)'
+//       />
+//     </clipPath>
+//   </defs>
+// </svg>;
